Initialize jekyll site
This is a really simple jekyll site. Basically it's jekyll-base in its minimal
working state.

The next step will be enabling multilingual content. It's one of the core
features of this theme and the one I'm more excited about. I don't know exactly
how I'm going to implement it.

Let's see how this adventure unfolds!