Adapt README.md to v0.2.0
The main objectives for this release were the material design and the use of
different color palettes. Although this is partially done it's becoming too
difficult to use and it's not looking exactly as I expected. I'll try to use
different layouts from this point on and see if this situation improves.