Load link contents with AJAX and update the rendered page with it
This is something I wanted to do at the end, or at least not right now. But I
came across https://github.com/joelhans/Jekyll-AJAX by @joelhans and it was easy
to adapt it to colos. That project doesn't have a LICENSE file
(joelhans/Jekyll-AJAX#2), but since that project has another contributor and
seeing the contents of the README I suppose it's absent by mistake. So I used
the code as it were MIT compatible.

> From the README: The best way to get functionality out-of-the-box is to clone
> this repo out of Github and use it as a starting point for your own creation.
> You can change the formatting, the types of layouts — whatever you want.